+++ libheif/codecs/uncompressed/unc_encoder_component_interleave.h
@@ -40,7 +40,7 @@ public:
 private:
   struct channel_component
   {
-    uint16_t component_idx;
+    uint32_t component_idx;
     heif_channel channel;
     heif_uncompressed_component_type component_type;
     heif_uncompressed_component_format component_format;
